requirements.txt
asyncio
PyYAML
psycopg2
sqlalchemy
jupyter